% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{assign_argument_defaults_addin}
\alias{assign_argument_defaults_addin}
\title{Assign argument defaults}
\usage{
assign_argument_defaults_addin()
}
\description{
Evaluates all comma-separated \code{x = y} expressions in a selection. The
intended usage is for selecting all of the text between the parentheses in a
\verb{function(text = "to", select = "and", evaluate)} call so that default
arguments values can be assigned to a user's global environment.
}
