% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{flip_backslashes_addin}
\alias{flip_backslashes_addin}
\title{Convert back slashes to forward slashes}
\usage{
flip_backslashes_addin()
}
\description{
Call this addin to convert \verb{\\\\drive\\folder} to \verb{//drive/folder}
}
