% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{wrap_rmd_addin}
\alias{wrap_rmd_addin}
\title{Wrap text but don't insert lines breaks into inline R code}
\usage{
wrap_rmd_addin()
}
\description{
Call this addin to wrap paragraphs in an R Markdown document.
}
\details{
The maximum line width can be set via \code{options(WrapRmd.width)}.
Further fine tuning can be done via options \code{WrapRmd.smart} and
\code{WrapRmd.extensions}, which are passed down to
\code{\link[commonmark:commonmark]{commonmark::markdown_commonmark()}}.
}
