% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{knit_selection_addin}
\alias{knit_selection_addin}
\title{Run selection through knitr and commonmark}
\usage{
knit_selection_addin()
}
\description{
Call this addin to preview output of an R markdown selection.
}
