% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapRmd-package.r
\name{wrap_rmd_addin}
\alias{wrap_rmd_addin}
\title{Wrap text but don't insert lines breaks into inline R code}
\usage{
wrap_rmd_addin()
}
\description{
Call this addin to wrap paragraphs in an R Markdown document
}

