% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{WrapRmd}
\alias{WrapRmd}
\title{WrapRmd}
\description{
Provides an RStudio addin for wrapping paragraphs in an RMarkdown document
without inserting linebreaks into spans of inline R code.
}
